refactor: 重构数据实体与数据访问层，统一基金相关字段命名
- 将 Fund、FundAdj、FundNav、FundMarket 实体中的 symbol 字段统一更名为 code，以保持命名一致性
- 删除各模块中重复的实体类、Mapper 接口和 Dao 类，统一使用 trader-base 模块中的定义
- 更新 MyBatis-Plus 配置，将 type-aliases-package 指向统一的基础实体包路径
- 修改相关服务类，更新字段引用以适配新的字段名
- 调整 Correlation 实体字段命名，将 symbol1/symbol2 改为 asset1/asset2，并修改主键类型
- 为 TaskLog 实体添加 remark 字段，增强日志记录能力

+++ trader-executor/src/main/java/cc/riskswap/trader/executor/service/InvestmentService.java
@@ -9,9 +9,9 @@ import org.springframework.stereotype.Service;
 
 import cc.riskswap.trader.executor.common.enums.InvestmentStatusEnum;
 import cc.riskswap.trader.executor.common.model.dto.InvestmentInfoDto;
-import cc.riskswap.trader.executor.dao.InvestmentDao;
-import cc.riskswap.trader.executor.dao.entity.Investment;
 import cc.riskswap.trader.executor.lock.InvestmentLock;
+import cc.riskswap.trader.base.dao.InvestmentDao;
+import cc.riskswap.trader.base.dao.entity.Investment;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
@@ -51,7 +51,7 @@ public class InvestmentService {
         if (investment == null) {
             return;
         }
-        Long investmentId = investment.getId();
+        Integer investmentId = investment.getId();
         investmentDao.updateStatusAndExecutorId(investmentId, InvestmentStatusEnum.RUNNING.code, executorId);
         log.info("Investment status updated to RUNNING: group={}, id={}", investment.getGroupName(), investmentId);
     }
@@ -63,7 +63,7 @@ public class InvestmentService {
         if (investment == null) {
             return;
         }
-        Long investmentId = investment.getId();
+        Integer investmentId = investment.getId();
         investmentDao.updateStatusAndExecutorId(investmentId, InvestmentStatusEnum.STOPPED.code, executorId);
         log.info("Investment status updated to STOPPED: group={}, id={}", investment.getGroupName(), investmentId);
     }
@@ -71,12 +71,12 @@ public class InvestmentService {
     /**
      * 删除任务
      */
-    public void deleteInvestment(Long investmentId) {
+    public void deleteInvestment(Integer investmentId) {
         investmentDao.removeById(investmentId);
         log.info("Investment deleted: id={}", investmentId);
     }
 
-    public void execute(Long investmentId) {
+    public void execute(Integer investmentId) {
         Investment investment = investmentDao.getById(investmentId);
         if (investment == null) {
             return;
@@ -84,7 +84,7 @@ public class InvestmentService {
         log.info("executing investment {}", investment.getGroupName());
     }
 
-    public void trigger(Long investmentId) {
+    public void trigger(Integer investmentId) {
         Investment investment = investmentDao.getById(investmentId);
         if (investment == null) {
             return;
@@ -92,7 +92,7 @@ public class InvestmentService {
         log.info("triggering investment {}", investment.getGroupName());
     }
 
-    public void init(Long investmentId) {
+    public void init(Integer investmentId) {
         Investment investment = investmentDao.getById(investmentId);
         if (investment == null) {
             return;
